requirements.txt
# 3.11.0

# init packages
fastapi
uvicorn
pyngrok
python-dotenv
kafka-python
pyspark
SQLAlchemy
psycopg2


# # pip freeze
# annotated-types==0.7.0
# anyio==4.8.0
# click==8.1.8
# colorama==0.4.6
# fastapi==0.115.6
# h11==0.14.0
# idna==3.10
# pydantic==2.10.4
# pydantic_core==2.27.2
# pyngrok==7.2.2
# python-dotenv==1.0.1
# PyYAML==6.0.2
# sniffio==1.3.1
# starlette==0.41.3
# typing_extensions==4.12.2
# uvicorn==0.34.0